wip: Merges Constraints into EndpointConstraints.
