--- src/main/java/org/jitsi/videobridge/Constraints.java
@@ -1,26 +1,0 @@
-package org.jitsi.videobridge;
-
-public class Constraints
-{
-    private final int idealHeight;
-
-    public Constraints(int idealHeight)
-    {
-        this.idealHeight = idealHeight;
-    }
-
-    public int getIdealHeight()
-    {
-        return idealHeight;
-    }
-
-    static Constraints makeMaxHeightEndpointConstraints(int idealHeight)
-    {
-        return new Constraints(idealHeight);
-    }
-
-    public EndpointConstraints of(String id)
-    {
-        return new EndpointConstraints(id, idealHeight);
-    }
-}
